requirements.txt
discord.py
flask
asyncio
